+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 * Fix source sets with a capital letter in the name (like integrationTest) support (#134)
 * Checkstyle:
   - Add support for checkstyle 13 (requiring java 21): checkstyle plugin will be disabled on lower java versions.
-  - Add quality.suppressCheckstyleRules option to be able to remove rules from the default xml config (automatic config modification)   
+  - Add quality.suppressCheckstyleRules option to be able to remove rules from the default xml config (automatic config modification)
+  - When auto fallback is enabled, incompatible rules disabled automatically (using new suppression option)
 
 ### 6.0.1 (2025-09-03)
 * Always add spotbugs-annotations dependency when plugin is available in classpath to